Fix ChainSyncer _test() function so that we can run it directly again
It was broken during the move from p2p/ to trinity/
